fix helpa list-panel scroll position (frameDot -1 zero-extended through untyped dispatch)
FossilOrigin-Name: 0601dae0693809afa24a3bf58b5c3da6230084b77d5e870c8379d244a8345216

+++ src/atk/text/textv.c
@@ -452,7 +452,7 @@ void textview__SetDataObject(self, dataObject)
 
     self->dot = text_CreateMark((struct text *) dataObject, 0, 0);
     self->top = text_CreateMark((struct text *) dataObject, 0, 0);
-    self->frameDot = text_CreateMark((struct text *) dataObject, -1, 0);
+    self->frameDot = text_CreateMark((struct text *) dataObject, (long)-1, 0);
     /* The frameDot must start at -1 because that is later interpreted
           (near the end of DoUpdate) as meaning that it has never been set */
     self->atMarker = text_CreateMark((struct text *) dataObject, 0, 0);
